Split Base64 and Url building out of 'utils'
