+++ src/Auth0.Core/UrlBuilding.cs
@@ -5,35 +5,8 @@ using System.Text;
 
 namespace Auth0.Core.Http
 {
-    internal static class Utils
+    internal static class UrlBuilder
     {
-        internal static byte[] Base64UrlDecode(string input)
-        {
-            var output = input;
-            output = output.Replace('-', '+');          // 62nd char of encoding
-            output = output.Replace('_', '/');          // 63rd char of encoding
-
-            switch (output.Length % 4)                  // Pad with trailing '='s
-            {
-                case 0: break;                          // No pad chars in this case
-                case 2: output += "=="; break;          // Two pad chars
-                case 3: output += "="; break;           // One pad char
-                default: throw new InvalidOperationException("Illegal base64url string!");
-            }
-
-            return Convert.FromBase64String(output);    // Standard base64 decoder
-        }
-
-        internal static string Base64UrlEncode(byte[] input)
-        {
-            var output = Convert.ToBase64String(input);
-            output = output.Replace('-', '+');
-            output = output.Replace('_', '/');
-            output = output.PadRight(output.Length + (4 - output.Length % 4) % 4, '=');
-
-            return output;
-        }
-
         internal static Uri BuildUri(string baseUrl, string resource, IDictionary<string, string> urlSegments, IDictionary<string, string> queryStrings, bool includeEmptyParameters = false)
         {
             // Replace the URL Segments
@@ -86,7 +59,7 @@ namespace Auth0.Core.Http
         /// See http://stackoverflow.com/a/6704287
         /// </summary>
         /// <param name="uriParts">The URI parts to combine.</param>
-        internal static string CombineUriParts(params string[] uriParts)
+        private static string CombineUriParts(params string[] uriParts)
         {
             var uri = string.Empty;
             if (uriParts != null && uriParts.Any())
@@ -101,6 +74,5 @@ namespace Auth0.Core.Http
             }
             return uri;
         }
-
     }
 }
